requirements.txt
ipython>=8.10
jupyterlab>=3.0
notebook
ipykernel
pyarrow

kedro>=0.19,<0.20
kedro-datasets[pandas-csvdataset,pandas-exceldataset,pandas-parquetdataset,plotly-plotlydataset,plotly-jsondataset,matplotlib-matplotlibwriter]>=3.0,<4.0
kedro-viz>=11.0.0


pandas
scikit-learn~=1.5.1
seaborn~=0.12.1
matplotlib
umap-learn
pyod
mlxtend
dvc[s3]